Enhanced code that finds radio serial ports. Added error reporting to the user if serial port isn't opening. Added command-line arguments for serial port (example: "-p /dev/ttyUSB0" or "--port /dev/ttyUSB0"). Cleaned up some of the debug messages. wfview also will now tell the user what radio was found in the status bar of the main window.
Added the begininning of code to determine the waveform type (center or
fixed). rigcommander now knows which is which. The next step is to emit
a signal when this changes, which then causes the main window to check
or uncheck the box (without triggering any other code).

Added some TODO remarks on mode (DV, DD).

Added some command line arguments that are not used at this time for --
host and --civ.